+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Integration with the [Confluent Schema Registry]
 - `kafkactl consume`: Support decoding Avro messages
 - `kafkactl get message`: Support decoding Avro messages
 - `kafkactl get topic`: Treat topics with single `_` prefix as internal (instead of double `_` prefix)
@@ -16,3 +17,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://github.com/fgrosse/kafkactl/compare/v1.0.0...HEAD
 [v1.0.0]: https://github.com/fgrosse/kafkactl/releases/tag/v1.0.0
 
+[schema-registry]: https://docs.confluent.io/platform/current/schema-registry/index.html